CI: Build and test CBC-optimizer in separate pipeline (#47)
